requirements.txt
aiomysql==0.2.0
fastapi==0.111.1
pillow==10.4.0
python-dotenv==1.0.1
python-jose==3.3.0
uvicorn==0.30.1
